[ 김하연 ] 기능 수정: 검색 엔드포인트 통합
