+++ routes/search.js
@@ -1,9 +1,9 @@
 const express = require('express');
-const { searchPosts } = require('../controller/main');
+const { searchPosts } = require('../controller/search');
 const { authUtil } = require('../middlewares/auth');
 
 const router = express.Router();
 
-router.get('/search', authUtil.isSignedIn, authUtil.isAuthorized, searchPosts );
+router.get('/',  searchPosts ); //authUtil.isSignedIn, authUtil.isAuthorized
 
 module.exports = router;
